move stuff up from ViewModel to View
