+++ kotlin/android/app/src/main/java/dev/firezone/android/features/session/ui/SessionViewModel.kt
@@ -21,15 +21,6 @@ internal class SessionViewModel
         internal lateinit var repo: Repository
         private val _serviceStatusLiveData = MutableLiveData<State>()
         private val _resourcesLiveData = MutableLiveData<List<Resource>>(emptyList())
-        private var showOnlyFavorites: Boolean = false
-
-        init {
-            viewModelScope.launch {
-                repo.favorites.collect { _ -> if (forceAllResourcesTab()) {
-                    showOnlyFavorites = false
-                } }
-            }
-        }
 
         val serviceStatusLiveData: MutableLiveData<State>
             get() = _serviceStatusLiveData
@@ -43,51 +34,7 @@ internal class SessionViewModel
 
         fun clearToken() = repo.clearToken()
 
-        // The subset of Resources to actually render
-        fun resourcesList(isInternetResourceEnabled: ResourceState): List<ResourceViewModel> {
-            val resources =
-                resourcesLiveData.value!!.map {
-                    if (it.isInternetResource()) {
-                        ResourceViewModel(it, isInternetResourceEnabled)
-                    } else {
-                        ResourceViewModel(it, ResourceState.ENABLED)
-                    }
-                }
-
-            return if (repo.favorites.value.isEmpty()) {
-                resources
-            } else if (showOnlyFavorites) {
-                resources.filter { repo.favorites.value.contains(it.id) }
-            } else {
-                resources
-            }
-        }
-
-        fun forceAllResourcesTab(): Boolean {
-            return repo.favorites.value.isEmpty()
-        }
-
         fun showFavoritesTab(): Boolean {
             return repo.favorites.value.isNotEmpty()
-        }
-
-        fun tabSelected(position: Int) {
-            showOnlyFavorites =
-                when (position) {
-                    RESOURCES_TAB_FAVORITES -> {
-                        true
-                    }
-
-                    RESOURCES_TAB_ALL -> {
-                        false
-                    }
-
-                    else -> throw IllegalArgumentException("Invalid tab position: $position")
-                }
-        }
-
-        companion object {
-            const val RESOURCES_TAB_FAVORITES = 0
-            const val RESOURCES_TAB_ALL = 1
         }
     }
